feat: connection redirect to editor panel
